AscendMemory/pyproject.toml
[project]
name = "ascend-memory"
version = "0.1.0"
description = "AscendAI Memory Service (MCP + REST)"
requires-python = ">=3.11"
dependencies = [
    "fastapi==0.128.0",
    "uvicorn==0.40.0",
    "mem0ai==1.0.3",
    "fastmcp==2.14.5",
    "pydantic==2.11.10",
    "pydantic-settings==2.12.0",
    "httpx==0.28.1",
    "python-multipart==0.0.20",
    "qdrant-client==1.16.2",
    "colorlog==6.10.1"
]

[project.optional-dependencies]
dev = [
    "pytest==9.0.2",
    "pytest-asyncio==1.3.0",
    "pytest-cov==7.0.0",
    "httpx==0.28.1",
    "asgi-lifespan==2.1.0",
    "black==26.1.0",
    "isort==7.0.0"
]

[tool.setuptools]
packages = ["src"]
